Move expressions types to own files
