+++ nikitov.dmitriy/S2/convert_expression.hpp
@@ -2,7 +2,7 @@
 #define CONVERT_EXPRESSION_HPP
 
 #include "queue.hpp"
-#include "expression_types.hpp"
+#include "infix_type.hpp"
 #include "postfix_expression.hpp"
 
 namespace nikitov
