+++ nikitov.dmitriy/S2/expression_types.hpp
@@ -33,68 +33,5 @@ namespace nikitov
     operand,
     operation
   };
-
-  struct InfixType
-  {
-  public:
-    InfixType() = default;
-    explicit InfixType(bool value);
-    explicit InfixType(long long value);
-    explicit InfixType(char value);
-
-    bool& getBracket();
-    long long& getOperand();
-    char& getOperation();
-    ExprTypeName getType() const;
-
-  private:
-    union
-    {
-      detail::Bracket bracket;
-      detail::Operand operand;
-      detail::Operation operation;
-    };
-    ExprTypeName type;
-  };
-
-  struct StackType
-  {
-  public:
-    StackType() = default;
-    explicit StackType(bool value);
-    explicit StackType(char value);
-
-    bool& getBracket();
-    char& getOperation();
-    ExprTypeName getType() const;
-
-  private:
-    union
-    {
-      detail::Bracket bracket;
-      detail::Operation operation;
-    };
-    ExprTypeName type;
-  };
-
-  struct PostfixType
-  {
-  public:
-    PostfixType() = default;
-    explicit PostfixType(long long value);
-    explicit PostfixType(char value);
-
-    long long& getOperand();
-    char& getOperation();
-    ExprTypeName getType() const;
-
-  private:
-    union
-    {
-      detail::Operand operand;
-      detail::Operation operation;
-    };
-    ExprTypeName type;
-  };
 }
 #endif
