+++ nikitov.dmitriy/S2/main.cpp
@@ -2,7 +2,7 @@
 #include <fstream>
 #include "queue.hpp"
 #include "stack.hpp"
-#include "expression_types.hpp"
+#include "infix_type.hpp"
 #include "postfix_expression.hpp"
 #include "input_expressions.hpp"
 #include "convert_expression.hpp"
