+++ nikitov.dmitriy/S2/postfix_type.hpp
@@ -1,0 +1,28 @@
+#ifndef POSTFIX_TYPE_HPP
+#define POSTFIX_TYPE_HPP
+
+#include "expression_types.hpp"
+
+namespace nikitov
+{
+  struct PostfixType
+  {
+  public:
+    PostfixType() = default;
+    explicit PostfixType(long long value);
+    explicit PostfixType(char value);
+
+    long long& getOperand();
+    char& getOperation();
+    ExprTypeName getType() const;
+
+  private:
+    union
+    {
+      detail::Operand operand;
+      detail::Operation operation;
+    };
+    ExprTypeName type;
+  };
+}
+#endif
